+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [data][diaggui] Added methods Diaggui.tf(), Diaggui.csd(), Diaggui.psd(), and
+  Diaggui.coh(). Also added attributes Diaggui.items, Diaggui.path.
+- [data][diaggui] Added Diaggui class for handling diaggui XML files.
+  The class is declared with the path to the file.
 - Using mypythonlibrary(https://github.com/terrencetec/mypythonlibrary) as
   a template to create this library.
 - This CHANGELOG file to hopefully serve as an evolving example of a